lab8/requirements.txt
tensorflow
numpy
Pillow
pillow
opencv-python
scipy